% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheetah_proxy.R
\name{add_row}
\alias{add_row}
\title{Add a row to a cheetah table}
\usage{
add_row(proxy, data)
}
\arguments{
\item{proxy}{A proxy object created by \code{cheetah_proxy()}.}

\item{data}{A single row dataframe for the new row in the table}
}
\description{
Adds a new row to a cheetah grid widget without redrawing the entire widget.
}
