% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheetah.R
\name{cheetah}
\alias{cheetah}
\title{Cheetah}
\usage{
cheetah(data, columns = NULL, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{data}{Data}

\item{columns}{Columns}

\item{width}{Width}

\item{height}{Height}

\item{elementId}{Element ID}
}
\description{
Some description.
}
