% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheetah_proxy.R
\name{cheetah_proxy}
\alias{cheetah_proxy}
\title{Cheetah Grid Proxy}
\usage{
cheetah_proxy(outputId, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{outputId}{The id of the cheetah table to create a proxy for.}

\item{session}{The Shiny session object. The default is to get the current session.}
}
\value{
A proxy object that can be used to modify the grid.
}
\description{
Creates a proxy object for a cheetah grid widget that can be used to modify the grid
without redrawing the entire widget.
}
