% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheetah_proxy.R
\name{delete_row}
\alias{delete_row}
\title{Delete a row from a cheetah grid}
\usage{
delete_row(proxy, row_index)
}
\arguments{
\item{proxy}{A proxy object created by \code{cheetah_proxy()}.}

\item{row_index}{A numeric value representing the index of the row to delete.}
}
\description{
Deletes a row from a cheetah grid widget without redrawing the entire widget.
}
