% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_ifelse.R
\name{js_ifelse}
\alias{js_ifelse}
\title{Convert an R-style conditional to a JS ternary expression}
\usage{
js_ifelse(condition, true_result, false_result)
}
\arguments{
\item{condition}{An R expression (logical) using \\%in\\%, \\%notin\\% or grepl(), etc.}

\item{true_result}{A string value to return when the condition is TRUE.}

\item{false_result}{A string value to return when the condition is FALSE.}
}
\value{
A character string containing a JavaScript ternary expression.
}
\description{
Convert an R-style conditional to a JS ternary expression
}
