% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheetah_utils.R
\name{number_format}
\alias{number_format}
\title{Numeric Column formatter}
\usage{
number_format(
  style = c("decimal", "currency", "percent", "unit"),
  currency = NULL,
  currency_display = c("symbol", "code", "narrowSymbol", "name"),
  currency_sign = c("standard", "accounting"),
  unit = NULL,
  unit_display = c("short", "narrow", "long"),
  locales = NULL,
  locale_options = NULL,
  digit_options = NULL,
  other_options = NULL
)
}
\arguments{
\item{style}{The formatting style to use. Must be one of the following:
\itemize{
\item \code{"decimal"} for plain number formatting action columns. Default.
\item \code{"currency"} for currency formatting
\item \code{"percent"} for percent formatting.
\item \code{"unit"} for unit formatting.
}}

\item{currency}{The ISO 4217 currency code to use for currency formatting. Must be provided if \code{style} is \code{"currency"}.}

\item{currency_display}{The display format for the currency. Must be one of the following:
\itemize{
\item \code{"symbol"} for the currency symbol. Default. Use the localized currency symbol e.g. $ for USD.
\item \code{"code"} for the currency code. Use the ISO 4217 currency code e.g. USD for US Dollar.
\item \code{"narrowSymbol"} for the narrow currency symbol. Use the narrow currency symbol e.g. \verb{$100} instead of \verb{$USD100}.
\item \code{"name"} for the currency name. Use the localized currency name e.g. Dollar for USD.
}}

\item{currency_sign}{The sign to use for the currency. Must be one of the following:
\itemize{
\item \code{"standard"} for the standard format. Default.
\item \code{"accounting"} for the accounting format. Use the accounting sign e.g. $100 instead of $USD100.
}}

\item{unit}{The unit to use for the unit formatting. Must be provided if \code{style} is \code{"unit"}.}

\item{unit_display}{The display format for the unit. Must be one of the following:
\itemize{
\item \code{"short"} for the short format. Default. E.g., 16 l
\item \code{"narrow"} for the narrow format. E.g., 16l
\item \code{"long"} for the long format. E.g., 16 liters
}}

\item{locales}{A character vector of BCP 47 language tags (e.g. \code{"en-US"} for English,
\code{"ja-JP"} for Japanese) specifying the locales to use for number formatting.
If NULL, the runtime's default locale is used. See
\href{https://learn.microsoft.com/en-us/openspecs/office_standards/ms-oe376/6c085406-a698-4e12-9d4d-c3b0ee3dbc4a}{BCP 47 language tags}
for reference.}

\item{locale_options}{A named list of options to customize the locale.}

\item{digit_options}{A named list of options to customize the digit.}

\item{other_options}{A named list of other options to customize the number formatting.
@note
Further details on customizing numeric formatting can be found in the
\href{https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Intl/NumberFormat/NumberFormat#parameters}{Intl.NumberFormat documentation}}
}
\value{
A list containing number formatting options that can be used to format numeric data in a column.
}
\description{
To add data formatting to a numeric type column.
}
\examples{
data <- data.frame(
  price_USD = c(125000.75, 299.99, 7890.45),
  price_EUR = c(410.25, 18750.60, 1589342.80),
  price_INR = c(2200.50, 134999.99, 945.75),
  price_NGN = c(120000, 2100045, 1750),
  liter = c(20, 35, 42),
  percent = c(0.875, 0.642, 0.238)
)

cheetah(
  data,
  columns = list(
    price_USD = column_def(
      name = "USD",
      column_type = number_format(
        style = "currency",
        currency = "USD"
      )
    ),
    price_EUR = column_def(
      name = "EUR",
      column_type = number_format(
        style = "currency",
        currency = "EUR",
        locales = "de-DE"
      )
    ),
    price_INR = column_def(
      name = "INR",
      column_type = number_format(
        style = "currency",
        currency = "INR",
        locales = "hi-IN"
      )
    ),
    price_NGN = column_def(
      name = "NGN",
      column_type = number_format(
        style = "currency",
        currency = "NGN"
      )
    ),
    liter = column_def(
      name = "Liter",
      column_type = number_format(
        style = "unit",
        unit = "liter",
        unit_display = "long"
      )
    ),
    percent = column_def(
      name = "Percent",
      column_type = number_format(style = "percent")
    )
  )
)

}
