% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheetah.R
\name{cheetah}
\alias{cheetah}
\title{Create a Cheetah Grid widget}
\usage{
cheetah(
  data,
  columns = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL,
  rownames = TRUE
)
}
\arguments{
\item{data}{A data frame or matrix to display}

\item{columns}{A list of column definitions. Each column can be customized using
\code{column_def()}.}

\item{width}{Width of the widget}

\item{height}{Height of the widget}

\item{elementId}{The element ID for the widget}

\item{rownames}{Logical. Whether to show rownames. Defaults to TRUE.}
}
\value{
An HTML widget object of class 'cheetah' that can be:
\itemize{
\item Rendered in R Markdown documents
\item Used in Shiny applications
\item Displayed in R interactive sessions
}
The widget renders as an HTML table with all specified customizations.
}
\description{
Creates a high-performance table widget using Cheetah Grid.
}
