% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheetah_utils.R
\name{column_def}
\alias{column_def}
\title{Column definition utility}
\usage{
column_def(
  name = NULL,
  width = NULL,
  min_width = NULL,
  max_width = NULL,
  column_type = NULL,
  action = NULL,
  style = NULL
)
}
\arguments{
\item{name}{Custom name.}

\item{width}{Column width.}

\item{min_width}{Column minimal width.}

\item{max_width}{Column max width.}

\item{column_type}{Column type. TBD}

\item{action}{Column action. TBD}

\item{style}{Column style.}
}
\value{
A list of column options to pass to the JavaScript API.
}
\description{
Needed by \link{cheetah} to customize
columns individually.
}
