% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheetah_utils.R
\name{column_def}
\alias{column_def}
\title{Column definition utility}
\usage{
column_def(
  name = NULL,
  width = NULL,
  min_width = NULL,
  max_width = NULL,
  column_type = NULL,
  action = NULL,
  menu_options = NULL,
  style = NULL
)
}
\arguments{
\item{name}{Custom name.}

\item{width}{Column width.}

\item{min_width}{Column minimal width.}

\item{max_width}{Column max width.}

\item{column_type}{Column type. By default, the column type is inferred from the data type of the column.
There are 7 possible options:
\itemize{
\item \code{"text"} for text columns.
\item \code{"number"} for numeric columns.
\item \code{"check"} for check columns.
\item \code{"image"} for image columns.
\item \code{"radio"} for radio columns.
\item \code{"multilinetext"} for multiline text columns.
\item \code{"menu"} for menu selection columns. If \code{column_type == "menu"},
action parameter must be set to "inline_menu" and menu_options must be provided.
Note: Works efficiently only in shiny.
}}

\item{action}{The action property defines column actions. Select
the appropriate Action class for the column type.
\itemize{
\item \code{"input"} for input action columns.
\item \code{"check"} for check action columns.
\item \code{"radio"} for radio action columns.
\item \code{"inline_menu"} for menu selection columns.
}}

\item{menu_options}{A list of menu options when using \code{column_type = "menu"}.
Each option should be a list with \code{value} and \code{label} elements.
The menu options must be a list of lists, each containing a \code{value}
and \code{label} element.
The \code{label} element is the label that will be displayed in the menu.}

\item{style}{Column style.}
}
\value{
A list of column options to pass to the JavaScript API.
}
\description{
Needed by \link{cheetah} to customize
columns individually.
}
