% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheetah_utils.R
\name{column_def}
\alias{column_def}
\title{Column definition utility}
\usage{
column_def(
  name = NULL,
  width = NULL,
  min_width = NULL,
  max_width = NULL,
  column_type = NULL,
  action = NULL,
  style = NULL
)
}
\arguments{
\item{name}{Custom name.}

\item{width}{Column width.}

\item{min_width}{Column minimal width.}

\item{max_width}{Column max width.}

\item{column_type}{Column type. There are 6 possible options:
\itemize{
\item \code{"text"} for text columns.
\item \code{"number"} for numeric columns.
\item \code{"check"} for check columns.
\item \code{"image"} for image columns.
\item \code{"radio"} for radio columns.
\item \code{"multilinetext"} for multiline text in columns.
}}

\item{action}{The action property defines column actions. Select
the appropriate Action class for the column type. For instance,
if the column type is \code{"text"}, the action can be \code{"input"}.
There are 3 supported actions:
\itemize{
\item \code{"input"} for input action columns.
\item \code{"check"} for check action columns.
\item \code{"radio"} for radio action columns.
}}

\item{style}{Column style.}
}
\value{
A list of column options to pass to the JavaScript API.
}
\description{
Needed by \link{cheetah} to customize
columns individually.
}
